separated view and fetcher logic
